Deprecate whole of sphinx.ext.mathbase
